[wip] abstracting OAuth
