+++ ksqldb-api-client/src/main/java/io/confluent/ksql/api/client/impl/ClientImpl.java
@@ -44,6 +44,7 @@ import io.confluent.ksql.security.AuthType;
 import io.confluent.ksql.security.Credentials;
 import io.confluent.ksql.security.CredentialsFactory;
 import io.confluent.ksql.security.KsqlClientConfig;
+import io.confluent.ksql.security.oauth.ClientSecretIdpConfig;
 import io.confluent.ksql.security.oauth.IdpConfig;
 import io.confluent.ksql.util.AppInfo;
 import io.confluent.ksql.util.KsqlConfig;
@@ -1000,23 +1001,13 @@ public class ClientImpl implements Client {
     }
 
     if (clientOptions.getAuthType() == AuthType.OAUTHBEARER) {
-      final IdpConfig idpConfig = clientOptions.getIdpConfig();
-      log.debug("Configuring bearer auth for clientId = {}",
-          idpConfig.getIdpClientId());
-      props.put(KsqlClientConfig.BEARER_AUTH_TOKEN_ENDPOINT_URL,
-          idpConfig.getIdpTokenEndpointUrl());
-      props.put(KsqlClientConfig.BEARER_AUTH_CLIENT_ID,
-          idpConfig.getIdpClientId());
-      props.put(KsqlClientConfig.BEARER_AUTH_CLIENT_SECRET,
-          idpConfig.getIdpClientSecret());
-      props.put(KsqlClientConfig.BEARER_AUTH_SCOPE,
-          idpConfig.getIdpScope());
-      props.put(KsqlClientConfig.BEARER_AUTH_SCOPE_CLAIM_NAME,
-          idpConfig.getIdpScopeClaimName());
-      props.put(KsqlClientConfig.BEARER_AUTH_SUB_CLAIM_NAME,
-          idpConfig.getIdpSubClaimName());
-      props.put(KsqlClientConfig.BEARER_AUTH_CACHE_EXPIRY_BUFFER_SECONDS,
-          idpConfig.getIdpCacheExpiryBufferSeconds());
+      IdpConfig idpConfig = clientOptions.getIdpConfig();
+      if (idpConfig instanceof ClientSecretIdpConfig) {
+        final ClientSecretIdpConfig clientSecretIdpConfig = (ClientSecretIdpConfig) clientOptions.getIdpConfig();
+        log.debug("Configuring bearer auth for clientId = {}",
+                clientSecretIdpConfig.getIdpClientId());
+      }
+      props.putAll(idpConfig.getIdpConfigs());
     }
 
     if (clientOptions.isUseTls()) {
